# swagger/sol_delilah.yaml
swagger: '2.0'
info:
  description: Delilah Restó is a food ordering app for restaurants
  version: 1.0.0
  title: Delilah Restó
  contact:
    email: solmartinezpisani@gmail.com
...
schemes:
  - https
  - http
paths:
...
  /cart/orders/{id}:
    get:
      tags:
        - cart
      summary: See shopping cart
      operationId: seeCart
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: id
          type: integer
          required: true
          description: Numeric ID of the order with a status of Shopping Cart.
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/SeeCartResponse'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
  /cart/orders/products/{idProduct}:
    post:
      tags:
        - cart
      summary: Add product to shopping cart
      operationId: addToCart
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: idProduct
          type: integer
          required: true
          description: Numeric ID of the product.
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/AddCartResponse'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            example: 'status_code: 404'
  /cart/orders/{idOrder}/products/{idProduct}:
    delete:
      tags:
        - cart
      summary: Delete product from shopping cart
      operationId: deleteFromCart
      produces:
        - application/json
      parameters:
        - name: User token
          in: header
          required: true
          type: string
          description: User JWT token (Bearer)
        - in: path
          name: idOrder
          type: integer
          required: true
          description: Numeric ID of the order to be deleted from the shopping cart.
        - in: path
          name: idProduct
          type: integer
          required: true
          description: Numeric ID of the product to be deleted from the shopping cart.
      responses:
        '200':
          description: Successful operation
          schema:
            $ref: '#/definitions/DeleteFromCartResponse'
        '400':
          description: Unsuccessful
          schema:
            $ref: '#/definitions/Authenticate'
        '404':
          description: Not Found
          schema:
            type: object
            properties:
              status_code:
                type: integer
                format: int64
                example: 404
...
definitions:
  Authenticate:
    type: object
    properties:
      message:
        type: string
        example: Header must contain JWT to verify identity.
...
  SeeCartResponse:
      type: array
      items:
        $ref: '#/definitions/Cart'
      example:
        - name: "Pizza"
          price: 500
          amount : 1
        - total_price: 500
        - address: "Cornelia Street"
  Cart: 
    type: object
    properties:
      name:
        type: string
        description: Pizza
      price:
        type: "integer"
        format: "int64"
        example: 500.00
      amount: 
        type: "integer"
        format: "int64"
        example: 1
      total_price:
        type: "integer"
        format: "int64"
        example: 500.00
      address:
        type: "string"
        example: "Cornelia Street"
  AddCartResponse:
    type: object
    properties:
      message:
        type: string
        example: Added to cart
  DeleteFromCartResponse:
    type: object
    properties:
      message:
        type: string
        example: Product removed from shopping cart
...